+++ endpoint/http/src/main/java/de/fraunhofer/iosb/ilt/faaast/service/endpoint/http/HttpEndpoint.java
@@ -79,9 +79,6 @@ public class HttpEndpoint implements Endpoint<HttpEndpointConfig> {
     }
 
 
-    /**
-     * Hide sever (Jetty) information in the http response header.
-     */
     private void configHttpResponseHeader() {
         HttpConfiguration httpConfig = new HttpConfiguration();
         httpConfig.setSendServerVersion(false);